refactor(view): split secrets dialog out of dialogs.rs (<1000 lines)
dialogs.rs was 1018 lines. Convert it to a dialogs/ directory module and move
the secrets-management dialog data (SecretInfoData, SecretsDialogData) into
dialogs/secrets.rs, re-exported via `pub use secrets::*` so existing
`dialogs::…` / `rustyclaw_view::…` paths are unchanged. mod.rs is now 783 lines.
